requirements.txt
Django==4.2.7
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
pytest-django==4.7.0
